+++ src/page-load.js
@@ -1,3 +1,4 @@
+import weatherApi from './weatherApi';
 const pageLoad = () => ({
     divContent : document.getElementById("content"),
     hTitle : document.createElement("h2"),
@@ -86,25 +87,10 @@ const pageLoad = () => ({
       await this.getWeatherData();
       console.log("hello world");
   },
-  getWeatherData(){
-    return new Promise(resolve => {
-        const that = this;
-    fetch(`http://api.openweathermap.org/data/2.5/weather?q=${this.nameCity.value}&units=${this.selectUnit.value}&callback=test&appid=a71219e79a6b01978ac3a9f3ffccca37`, {
-        method: 'get',
-        mode: 'cors'
-      })
-            .then(response => {
-                if (!response.ok) {
-                    throw Error(response.statusText);
-                }
-                return response.text();
-            }
-                )
-      .then(json => {  
-            let jsonChars = json.split("");
-            jsonChars = jsonChars.slice(5, jsonChars.length-1).join("");
-            let data = JSON.parse(jsonChars);
-            console.log(data);
+  async getWeatherData(){
+    const result = await weatherApi().callWeatherApi(this.nameCity.value, this.selectUnit.value);
+    if(result.status){
+            const data = result.response;
             this.paragraph.innerHTML=`<h3>${this.nameCity.value} weather condition</h3>
             <p><img src="http://openweathermap.org/img/wn/${data.weather[0].icon}@2x.png" width="100" height="100" />
             <label>${data.weather[0].description}</label>
@@ -116,19 +102,16 @@ const pageLoad = () => ({
             <p><strong>pressure:</strong> ${data.main.pressure} hPa</p>
             <p><strong>humidity:</strong> ${data.main.humidity} %</p>
             <p><strong>wind speed:</strong> ${data.wind.speed} km/h</p>`;
-            this.divContent.style=`background-image: url("../images/${data.weather[0].icon}.jpg");`;
+            document.getElementsByTagName("body")[0].style=`background-image: url("../images/${data.weather[0].icon}.jpg");`;
             this.nameCity.value="";
             this.selectUnit.value="metric";
             this.addLabelErrorColor(0);
             this.labelError.innerText="search successfully completed";
-            resolve("result");
-        })
-    .catch(function(err) {
-        that.addLabelErrorColor(2);
-        that.labelError.innerText=err;
-        resolve("result");
-      });
-    });
+    }
+    else{
+      this.addLabelErrorColor(2);
+      this.labelError.innerText=result.response;
+    }
   }
   ,
   addLabelErrorColor(index){
