+++ styles/styles.css
@@ -11,11 +11,7 @@
     opacity:0.8;
 }
 
-.content{
+body{
     background-repeat: repeat;
     background-size: auto 477px;
 }
-
-/* .background_image_02d{
-    background-image: url("../images/02d.jpg");
-} */
